requirements.txt
asgiref==3.7.2
black==23.3.0
click==8.1.3
Django==4.2.2
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
flake8==6.0.0
Markdown==3.4.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
